Use List of Keys for acted_in relation
This method has almost all advantages of a Relationship document,
but allows us to prevent a double query for at least one side of
the relation. Building indexes should also be a lot quicker. The
reason for that is that this reduces the number of documents by
approximately 90%, because we don't need a document for every
acted_in relation.

This method is slightly more optimized for reads, while a relationship
document is more optimized for lots of concurrent writes (this could
cause conflicts using a List of Keys in a distributed setting). Since
both movies and actors are mostly only updated when a new movie comes out
, updates should be relatively infrequent, making this a better method
for our use-case.